+++ packages/core/src/shard/geometry-util.js
@@ -1,0 +1,3 @@
+import inherits from 'inherits/inherits_browser.js'
+
+export { inherits }